Dismis-master/requirements.txt
beautifulsoup4==4.7.1
bitstring==3.1.5
certifi==2018.11.29
chardet==3.0.4
click==6.7
DateTime==4.3
facebook-sdk==3.1.0
facepy==1.0.9
feedparser==5.2.1
HoroscopeGenerator==0.1.8
html5lib==1.0.1
idna==2.8
inflect==2.1.0
jaraco.itertools==4.4.1
keyring==8.7
keyrings.alt==1.3
more-itertools==6.0.0
nltk==3.4
numpy==1.16.2
Pillow==5.4.1
pip==19.0.3
playsound==1.2.2
psutil==5.5.1
PyAudio==0.2.11
pygame==1.9.4
pyicloud==0.9.1
pypiwin32==223
pytesseract==0.2.6
pyttsx3==2.7
pytz==2018.9
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==3.13
requests==2.21.0
selenium==3.141.0
setuptools==40.8.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
SpeechRecognition==3.8.1
tzlocal==1.5.1
urllib3==1.24.1
webencodings==0.5.1
wikipedia==1.4.0
wolframalpha==3.0.1
xmltodict==0.12.0
zope.interface==4.6.0
